dags/data_to_google_test.py
import unittest
from unittest.mock import patch
import internal_unit_testing
from . import data_to_google

class TestStockDataDag(unittest.TestCase):

    def test_dag_import(self):
        internal_unit_testing.assert_has_valid_dag(data_to_google)

    def test_move_data_task(self):
        task = data_to_google.dag.get_task('move_data_from_sheet_to_gcs')
        self.assertIsNotNone(task)
        self.assertEqual(task.task_id, 'move_data_from_sheet_to_gcs')

    @patch("path_to_your_module_containing_the_function.storage.Client")
    @patch("path_to_your_module_containing_the_function.gspread.authorize")
    def test_move_data_function(self, mock_gspread_authorize, mock_storage_client):
        # Mock the behavior of gspread and GCS so no real operations are performed
        mock_gspread_authorize.return_value = ...  # Mocked return value
        mock_storage_client.return_value = ...  # Mocked return value

        from path_to_your_module_containing_the_function import move_data_from_sheet_to_gcs
        result = move_data_from_sheet_to_gcs()
        self.assertEqual(result, expected_result)  # Replace expected_result with what you expect

if __name__ == '__main__':
    unittest.main()
