requirements.txt
apache-airflow==2.3.3
yfinance>=0.2.1
gspread>=3.7.0
oauth2client>=4.1.3
google-cloud-storage>=1.42.0
